requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch==2.1.0
torchvision==0.16.0
Pillow==10.0.1
pycocotools==2.0.7
python-multipart==0.0.6